feat(frontend): one task workspace merging management and review views
